pyproject.toml
[tool.poetry]
name = "nextcloud-video-frame-extractor"
version = "0.1.0"
description = ""
authors = ["brennoflavio <contato@brennoflavio.com.br>"]
readme = "README.md"
packages = [{include = "nextcloud_video_frame_extractor"}]

[tool.poetry.dependencies]
python = "^3.10"
webdavclient3 = "^3.14.6"
black = "^23.9.1"
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
